requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
requests>=2.31.0
pytrends>=4.9.2
python-dotenv>=1.0.0
plotly>=5.18.0
openpyxl>=3.1.2
tqdm>=4.66.0
python-aliexpress-api>=2.0.0
schedule>=1.2.0
